requirements.txt
Django==2.2.6
djangorestframework==3.10.3
django-cors-headers==3.1.1
django-simple-history==2.8.0
django-filter==2.2.0
python-dateutil==2.8.2
psycopg2==2.8.4
requests==2.27.1
urllib3==1.26.8
pika==1.1.0
uwsgi==2.0.19
pylint==2.4.3
jsonschema==3.1.1
json-logic-qubit==0.9.1
pydash==4.7.6
gql==2.0.0
coverage==6.4.4
django-nose==1.4.7